+++ src/types/address_space/kernel_as.c
@@ -1,5 +1,5 @@
 /*
- * Copyright 2020 United States Government
+ * Copyright 2023 United States Government
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -191,7 +191,7 @@ static gboolean kernel_address_equal(const address *a, const address *b)
 static guint kernel_address_hash(const address *a)
 {
     struct kernel_address *ka = (struct kernel_address *)a;
-    return ka->kaddr;
+    return (guint) ka->kaddr;
 }
 
 static void *kernel_read_bytes(address *a, size_t size)
